+++ src/pages/HomePage/HomePage.js
@@ -1,8 +1,6 @@
 import { HomePageStyled } from "./HomePageStyled";
 import WeekTabs from "../../components/weekDayTime/weekTabs/WeekTabs";
 import WeekTabContent from "../../components/weekDayTime/weekTabContent/WeekTabContent";
-import CurrentWeekRange from "../../components/weekDayTime/currentWeekRange/CurrentWeekRange";
-import TaskToggle from "../../components/taskToggle/TaskToggle";
 
 import Footer from "../../components/footer/Footer";
 import { useState, useEffect } from "react";
@@ -47,13 +45,9 @@ const HomePage = () => {
       {state.width < 767 && <WeekTabs />}
 
       <div className="right-side">
-        <CurrentWeekRange />
-
         <WeekTabContent />
-        <TaskToggle />
 
         <Footer />
-        {/* пока футер только для десктопа, подожду адаптивку от Тани  :) */}
       </div>
     </HomePageStyled>
   );
